pyproject.toml
[tool.poetry]
name = "pythonjwtlearning"
version = "0.1.0"
description = ""
authors = ["thach.elsa <thach.nguyen@elsanow.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
